+++ CHANGELOG.md
@@ -324,6 +324,7 @@ that error is that the account doesn't exist.
 * [\#4843](https://github.com/cosmos/cosmos-sdk/issues/4843) Add RegisterEvidences function in the codec package to register
   Tendermint evidence types with a given codec.
 * (rest) [\#3867](https://github.com/cosmos/cosmos-sdk/issues/3867) Allow querying for genesis transaction when height query param is set to zero.
+* (rest) [\#5113](https://github.com/cosmos/cosmos-sdk/issues/5113) Add CreateValidatorHandlerFn and EditValidatorHandlerFn in REST.
 * [\#2020](https://github.com/cosmos/cosmos-sdk/issues/2020) New keys export/import command line utilities to export/import private keys in ASCII format
   that rely on Keybase's new underlying ExportPrivKey()/ImportPrivKey() API calls.
 * [\#3565](https://github.com/cosmos/cosmos-sdk/issues/3565) Implement parameter change proposal support.